PokeStats.js moved to src/data/PokeStats.js
